+++ packages/website/siteConfig.js
@@ -33,7 +33,7 @@ const siteConfig = {
   // For no header links in the top nav bar -> headerLinks: [],
   headerLinks: [
     { doc: 'guide-about', label: 'Docs' },
-    { href: 'https://github.com/node-serialport/node-serialport', label: 'GitHub' },
+    { href: 'https://github.com/serialport/node-serialport', label: 'GitHub' },
     // { page: 'help', label: 'Help' },
     // { blog: true, label: 'Blog' },
   ],
@@ -81,10 +81,10 @@ const siteConfig = {
 
   // You may provide arbitrary config keys to be used as needed by your
   // template. For example, if you need your repo's URL...
-  repoUrl: 'https://github.com/node-serialport/node-serialport',
-  editUrl: `https://github.com/node-serialport/node-serialport/edit/master/docs/`,
+  repoUrl: 'https://github.com/serialport/node-serialport',
+  editUrl: `https://github.com/serialport/website/edit/master/docs/`,
   scrollToTop: true,
-  translationRecruitingLink: 'https://crowdin.com/project/node-serialport',
+  // translationRecruitingLink: 'https://crowdin.com/project/node-serialport',
 }
 
 module.exports = siteConfig