+++ packages/website/CHANGELOG.md
@@ -8,9 +8,9 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 ### Bug Fixes
 
-* conflicting website npm script name ([f6a800e](https://github.com/node-serialport/node-serialport/commit/f6a800e))
+* conflicting website npm script name ([f6a800e](https://github.com/serialport/node-serialport/commit/f6a800e))
 
 
 ### Features
 
-* move cli tools to their own packages ([#1664](https://github.com/node-serialport/node-serialport/issues/1664)) ([103498e](https://github.com/node-serialport/node-serialport/commit/103498e)), closes [#1659](https://github.com/node-serialport/node-serialport/issues/1659)
+* move cli tools to their own packages ([#1664](https://github.com/serialport/node-serialport/issues/1664)) ([103498e](https://github.com/serialport/node-serialport/commit/103498e)), closes [#1659](https://github.com/serialport/node-serialport/issues/1659)